+++ src/org/soulspace/overarch/adapter/reader/model_reader.clj
@@ -2,7 +2,10 @@
   "Functions for reading and building models and setting the repository state.
    The multimethods should be implemented by specific readers."
   (:require [clojure.set :as set]
-            [org.soulspace.overarch.application.model-repository :as repo]))
+            [org.soulspace.overarch.application.model-repository :as repo]
+            [org.soulspace.overarch.util.functions :as fns]
+            [org.soulspace.overarch.domain.element :as el]
+            [org.soulspace.overarch.domain.model :as model]))
 
 ;;;
 ;;; Reader functions
@@ -27,6 +30,69 @@
 ;;;
 ;;; Model building functions
 ;;;
+(defn build-id->element
+  "Returns the id->element index."
+  [elements]
+  (->> elements
+       (map (fn [e] [(:id e) e]))
+       (into {})))
+
+(defn build-id->parent-id
+  "Returns the id->parent-id index."
+  [elements]
+  (->> elements
+       (filter el/model-relation?)
+       (filter #(= :contained-in (:el %)))
+       (map (fn [e] [(:from e) (:to e)]))
+       (into {})))
+
+(defn build-referrer-id-relations
+  "Returns the referrer-id-relations index"
+  [elements]
+  (->> elements
+       (filter el/model-relation?)
+       (group-by :from)))
+
+(defn build-referred-id-relations
+  "Returns the referred-id-relations index"
+  [elements]
+  (->> elements
+       (filter el/model-relation?)
+       (group-by :to)))
+
+(defn update-indices
+  "Returns a `model` with updated indices."
+  [model]
+  (assoc model
+         :id->element (build-id->element (model/elements model))
+         :id->parent-id (build-id->parent-id (model/elements model))
+         :referred-id->relations (build-referred-id-relations (model/elements model))
+         :referrer-id->relations (build-referrer-id-relations (model/elements model))))
+
+(comment
+  (def model-elements
+    #{{:el :system :id :a/a-system}
+      {:el :container :id :a.a/a-container}
+      {:el :system :id :b/b-system}
+      {:el :contained-in
+       :id :a.a/a-container-contained-in-a-system
+       :from :a.a/a-container
+       :to :a/a-system}
+      {:el :request
+       :id :a/a-system-calls-b-system
+       :from :a/a-system
+       :to :b/b-system}
+      {:el :request
+       :id :a.a/a-container-calls-b-system
+       :from :a.a/a-container
+       :to :b/b-system}})
+  (fns/data-tapper (build-id->element model-elements))
+  (fns/data-tapper (build-id->parent-id model-elements))
+  (fns/data-tapper (build-referrer-id-relations model-elements))
+  (fns/data-tapper (build-referred-id-relations model-elements))
+  ;
+  )
+
 (defn merge-model
   "Merges the `model` with the `other` model."
   ([]